replace step function continue 2
